requirements_full.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pyhd3eb1b0_1
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.9.3=pyha847dfd_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.1.19=h06a4308_1
certifi=2020.12.5=py38h06a4308_0
cffi=1.14.3=py38h261ae71_2
chardet=3.0.4=py38h06a4308_1003
conda=4.9.2=py38h06a4308_0
conda-build=3.21.4=py38h06a4308_0
conda-package-handling=1.7.2=py38h03888b9_0
coverage=5.5=py38h27cfd23_2
cryptography=3.2.1=py38h3c74f83_1
cudatoolkit=11.1.74=h6bb024c_0
cython=0.29.22=py38h2531618_0
decorator=4.4.2=pyhd3eb1b0_0
expect=5.45.4=h555a92e_0
ffmpeg=4.3=hf484d3e_0
filelock=3.0.12=pyhd3eb1b0_1
freetype=2.10.4=h5ab3b9f_0
gdown=3.12.2=pyh9f0ad1d_0
glob2=0.7=pyhd3eb1b0_0
gmp=6.2.1=h2531618_2
gnutls=3.6.5=h71b1129_1002
grpcio=1.36.1=py38h2157cd5_1
icu=58.2=he6710b0_3
idna=2.10=py_0
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
ipython=7.21.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py38_0
jinja2=2.11.3=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libarchive=3.4.2=h62408e4_0
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
liblief=0.10.1=he6710b0_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py38h06a4308_0
markupsafe=1.1.1=py38h7b6447c_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
ninja=1.10.2=py38hff7bd54_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openssl=1.1.1j=h27cfd23_0
parso=0.8.1=pyhd3eb1b0_0
patchelf=0.12=h2531618_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.1=py38he98fc37_0
pip=20.2.4=py38h06a4308_0
pkginfo=1.7.0=py38h06a4308_0
prompt-toolkit=3.0.8=py_0
protobuf=3.14.0=py38h2531618_1
psutil=5.8.0=py38h27cfd23_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-lief=0.10.1=py38h403a769_0
pycosat=0.6.3=py38h7b6447c_1
pycparser=2.20=py_2
pygments=2.8.0=pyhd3eb1b0_0
pyopenssl=19.1.0=pyhd3eb1b0_1
pysocks=1.7.1=py38h06a4308_0
python=3.8.8=hdb3f193_4
python-libarchive-c=2.9=pyhd3eb1b0_0
pytorch=1.8.0=py3.8_cuda11.1_cudnn8.0.5_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py38h27cfd23_1
readline=8.0=h7b6447c_0
requests=2.24.0=py_0
ripgrep=12.1.1=0
ruamel_yaml=0.15.87=py38h7b6447c_1
setuptools=50.3.1=py38h06a4308_1
six=1.15.0=py38h06a4308_0
snakeviz=2.0.1=py38_0
soupsieve=2.2=pyhd3eb1b0_0
sqlite=3.33.0=h62c20be_0
tensorboard=1.15.0=py38_0
tk=8.6.10=hbc83047_0
torchtext=0.9.0=py38
torchvision=0.9.0=py38_cu111
tornado=6.1=py38h27cfd23_0
tqdm=4.51.0=pyhd3eb1b0_0
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.25.11=py_0
wcwidth=0.2.5=py_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.35.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
